refactor: separate component
