+++ src/app/page.tsx
@@ -1,95 +1,12 @@
-'use client'
-
-import { useState } from 'react'
-import { Dialog, DialogPanel } from '@headlessui/react'
-import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline'
 import Image from 'next/image'
 import Link from 'next/link'
-
-const navigation = [
-  { name: '概要', href: '#' },
-  { name: '次回開催', href: '#' },
-  { name: '過去の開催', href: '#' },
-  { name: 'スタッフ', href: '#' },
-  { name: 'Shirankedo', href: '#' },
-]
+import Header from '@/components/Header'
 
 export default function Home() {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
-
-  const naniwaNogIcon = (
-    <Link href={"/"}>
-      <Image
-        src="/favicon.ico"
-        alt="naniwa nog icon"
-        width={48}
-        height={48}
-        className="-m-1.5"
-      />
-    </Link>
-  )
 
   return (
     <div className="bg-white">
-      <header className="fixed inset-x-0 pb-2 top-0 z-50 bg-white/70 backdrop-blur">
-        {/* For PC */}
-        <div className="mx-auto max-w-7xl">
-          <div className="px-6 pt-6 lg:max-w-2xl lg:pl-8 lg:pr-0">
-            <nav aria-label="Global" className="flex items-center justify-between lg:justify-start">
-              {naniwaNogIcon}
-              <button
-                type="button"
-                onClick={() => setMobileMenuOpen(true)}
-                className="-m-2.5 rounded-md p-2.5 text-gray-700 lg:hidden"
-              >
-                <span className="sr-only">Open main menu</span>
-                <Bars3Icon aria-hidden="true" className="size-6" />
-              </button>
-              <div className="hidden lg:ml-12 lg:flex lg:gap-x-14">
-                {navigation.map((item) => (
-                  <Link key={item.name} href={item.href} className="text-sm/6 font-semibold text-gray-900">
-                    {item.name}
-                  </Link>
-                ))}
-              </div>
-            </nav>
-          </div>
-        </div>
-
-        {/* For mobile */}
-        <Dialog open={mobileMenuOpen} onClose={setMobileMenuOpen} className="lg:hidden">
-          <div className="fixed inset-0 z-50" />
-          <DialogPanel className="fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
-            <div className="flex items-center justify-between">
-              {naniwaNogIcon}
-              <button
-                type="button"
-                onClick={() => setMobileMenuOpen(false)}
-                className="-m-2.5 rounded-md p-2.5 text-gray-700"
-              >
-                <span className="sr-only">Close menu</span>
-                <XMarkIcon aria-hidden="true" className="size-6" />
-              </button>
-            </div>
-            <div className="mt-6 flow-root">
-              <div className="-my-6 divide-y divide-gray-500/10">
-                <div className="space-y-2 py-6">
-                  {navigation.map((item) => (
-                    <Link
-                      key={item.name}
-                      href={item.href}
-                      className="-mx-3 block rounded-lg px-3 py-2 text-base/7 font-semibold text-gray-900 hover:bg-gray-50"
-                    >
-                      {item.name}
-                    </Link>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </DialogPanel>
-        </Dialog>
-
-      </header>
+      <Header />
 
       <div className="relative">
         <div className="mx-auto max-w-7xl">
